+++ packages/components/layout/examples/basic/basic.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { THY_SITE_SETTINGS } from '@tethys/pro/core';
 
 @Component({
     selector: 'thy-pro-layout-basic-example',
@@ -7,6 +6,29 @@ import { THY_SITE_SETTINGS } from '@tethys/pro/core';
     styleUrls: ['./basic.component.scss']
 })
 export class ThyProLayoutBasicExampleComponent implements OnInit {
+    menus = [
+        {
+            data: {
+                title: '菜单1',
+                icon: 'menu'
+            }
+        },
+        {
+            data: {
+                title: '菜单2',
+                icon: 'menu'
+            },
+            children: [
+                {
+                    data: {
+                        title: '菜单2',
+                        icon: 'menu'
+                    }
+                }
+            ]
+        }
+    ];
+
     constructor() {}
 
     ngOnInit(): void {}